Move Handler to handler package
