--- pkg/dnsserver/dnscontroller/lighthouse.go
@@ -1,9 +1,0 @@
-package dnscontroller
-
-import (
-	"github.com/submariner-io/lighthouse/pkg/multiclusterservice"
-)
-
-type Lighthouse struct {
-	MultiClusterServices *multiclusterservice.Map
-}
